Split up parse.rs into multiple files
